backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.36
pydantic>=2.10.6
pydantic-settings>=2.7.1
python-multipart>=0.0.20
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
cryptography>=44.0.0
python-docx>=1.1.2
pymupdf>=1.25.1
python-magic>=0.4.27
openai>=1.59.0
